feat(MultiPlatform): separated modules to ease true multi platform development
